Move all homology-related components into own directory.
