+++ src/uniprotkb/components/entry/homologs/AgrParalogy.tsx
