+++ src/uniprotkb/components/entry/homologs/getHomologyMethodColumnConfig.tsx
