+++ src/uniprotkb/components/entry/homologs/styles/agr-homology.module.scss
